DataPrep4LLM_Algos/requirements.txt
PyMuPDF==1.24.9
numpy==1.26.4
networkx==3.3
pillow==10.4.0
matplotlib==3.9.0
pycocotools==2.0.8
torch==2.3.1
torchvision==0.18.1
ultralytics==8.2.59
surya-ocr==0.4.15
rapid-table==0.1.3
rapidocr-paddle==1.3.26
paddlepaddle-gpu==2.6.1